added validation students

+++ src/main/java/com/ovsyanka/springbootreactaws/StudentService.java
@@ -2,6 +2,8 @@ package com.ovsyanka.springbootreactaws;
 
 import com.ovsyanka.springbootreactaws.student.Student;
 import com.ovsyanka.springbootreactaws.student.StudentRepository;
+import com.ovsyanka.springbootreactaws.student.exception.BadRequestException;
+import com.ovsyanka.springbootreactaws.student.exception.StudentNotFoundException;
 import jakarta.servlet.ServletContext;
 import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
@@ -22,12 +24,21 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
